+++ src/components/form/form.style.ts
@@ -45,7 +45,9 @@ export const StyledFormContent = styled.div<StyledFormContentProps>(
 );
 
 interface StyledFormFooterProps {
+  hasFooterChildren?: boolean;
   stickyFooter?: boolean;
+  stickyFooterVariant?: "light" | "grey";
   fullWidthButtons?: boolean;
   buttonAlignment?: FormButtonAlignment;
 }
@@ -53,10 +55,14 @@ interface StyledFormFooterProps {
 export const StyledFormFooter = styled.div.attrs(
   applyBaseTheme,
 )<StyledFormFooterProps>`
-  align-items: center;
-  display: flex;
-  flex-wrap: wrap;
-  gap: var(--sizing200);
+  ${({ hasFooterChildren }) =>
+    !hasFooterChildren &&
+    css`
+      align-items: center;
+      display: flex;
+      flex-wrap: wrap;
+      gap: var(--sizing200);
+    `}
 
   ${({ buttonAlignment }) =>
     buttonAlignment === "right" &&
@@ -64,7 +70,7 @@ export const StyledFormFooter = styled.div.attrs(
       justify-content: flex-end;
     `}
 
-  ${({ stickyFooter, fullWidthButtons }) => css`
+  ${({ stickyFooter, stickyFooterVariant, fullWidthButtons }) => css`
     ${!stickyFooter &&
     css`
       margin-top: 48px;
@@ -72,9 +78,13 @@ export const StyledFormFooter = styled.div.attrs(
 
     ${stickyFooter &&
     css`
-      background-color: var(--colorsUtilityYang100);
-      box-shadow: 0 -4px 12px 0 rgba(0, 0, 0, 0.05);
+      background-color: ${stickyFooterVariant === "light"
+        ? "var(--colorsUtilityYang100)"
+        : "var(--colorsUtilityMajor025)"};
+      box-shadow: var(--boxShadow150);
       box-sizing: border-box;
+      ${stickyFooterVariant === "grey" &&
+      "border-top: 1px solid var(--colorsUtilityMajor050);"}
       padding: 16px 32px;
       width: 100%;
       z-index: 1000;